CAP_SYS/CAP_SYS/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
mysql-connector-python==9.0.0
pandas==2.2.2
transformers==4.41.2
torch==2.2.2
sentence-transformers==2.2.2
bertopic==0.15.0
scikit-learn==1.4.2
umap-learn==0.5.6
hdbscan==0.8.33
nltk==3.8.1
